MOVE babel-loader_function to preview folder & ADD COPY of tailored for manager
The big change is that now, the babel presets & plugins used for storybook (react) code are ensured for manager injected code

This means addons referenced from preview.js & manager.js will load with babel config including react preset, etc.
