+++ lib/core/package.json
@@ -32,6 +32,7 @@
     "@babel/plugin-syntax-dynamic-import": "^7.2.0",
     "@babel/plugin-transform-react-constant-elements": "^7.2.0",
     "@babel/preset-env": "^7.4.5",
+    "@babel/preset-react": "^7.8.3",
     "@storybook/addons": "6.0.0-alpha.1",
     "@storybook/channel-postmessage": "6.0.0-alpha.1",
     "@storybook/client-api": "6.0.0-alpha.1",
@@ -45,6 +46,7 @@
     "airbnb-js-shims": "^2.2.1",
     "ansi-to-html": "^0.6.11",
     "autoprefixer": "^9.7.2",
+    "babel-merge": "^3.0.0",
     "babel-plugin-add-react-displayname": "^0.0.5",
     "babel-plugin-emotion": "^10.0.20",
     "babel-plugin-macros": "^2.7.0",
