+++ lib/core/src/server/manager/manager-webpack.config.js
@@ -13,7 +13,7 @@ import findCacheDir from 'find-cache-dir';
 import { version } from '../../../package.json';
 import { getManagerHeadHtml } from '../utils/template';
 import { loadEnv } from '../config/utils';
-import babelLoader from '../common/babel-loader';
+import babelLoader from './babel-loader-manager';
 
 const coreDirName = path.dirname(require.resolve('@storybook/core/package.json'));
 const context = path.join(coreDirName, '../../node_modules');
