+++ lib/core/src/server/preview/iframe-webpack.config.js
@@ -10,7 +10,7 @@ import VirtualModulePlugin from 'webpack-virtual-modules';
 
 import resolveFrom from 'resolve-from';
 
-import createBabelLoader from '../common/babel-loader';
+import createBabelLoader from './babel-loader-preview';
 import { nodeModulesPaths, loadEnv } from '../config/utils';
 import { getPreviewHeadHtml, getPreviewBodyHtml } from '../utils/template';
 import { toRequireContextString } from './to-require-context';
